requirements.txt
altgraph==0.17.4
apturl==0.5.2
asgiref==3.7.2
bcrypt==3.2.0
beautifulsoup4==4.10.0
black==24.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.6
command-not-found==0.3
coverage==7.4.2
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.2
docx2txt==0.8
duplicity==0.8.21
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==23.2.1
fasteners==0.14.1
flake8==7.0.0
future==0.18.2
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==2.0.0
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mccabe==0.7.0
meson==0.61.2
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==23.2
paramiko==2.9.3
pathspec==0.12.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
pluggy==1.4.0
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.11.1
pycups==2.0.1
pyflakes==3.2.0
Pygments==2.11.2
PyGObject==3.42.1
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==8.0.1
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-factoryboy==2.6.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.4
systemd-python==234
tomli==2.0.1
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
urwid==2.1.2
usb-creator==0.3.7
wadllib==1.3.6
webencodings==0.5.1
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
